pyproject.toml
[project]
name = "ai-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Danny Pulickal",email = "dannypulickal@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "streamlit (>=1.56.0,<2.0.0)",
    "langchain (>=1.2.15,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "langchain-openai (>=1.2.1,<2.0.0)",
    "langchain-core (>=1.3.2,<2.0.0)"
]

[tool.poetry]
packages = [{include = "ai_chatbot", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "ruff (>=0.15.12,<0.16.0)"
]

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]  # Errors, Pyflakes, Import sorting

[tool.ruff.format]
quote-style = "double"
